pyproject.toml
[project]
name = "prompt-optimizer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-groq>=0.3.6",
    "langgraph>=0.6.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]
